LayoutTests/fast/mediastream/RTCPeerConnection-stats-expected.txt

Tests the RTCPeerConnection stats interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc = new webkitRTCPeerConnection(null) did not throw exception.
PASS pc.getStats(statsHandler) did not throw exception.
PASS statsHandler was called
PASS status is non-null.
PASS result.length is 0
PASS successfullyParsed is true

TEST COMPLETE
